requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
groq
python-dotenv==1.2.2